Rename operations -> methods
