+++ tests_random/methods/add_range_column.py
