+++ tests_random/methods/cbind_self.py
