+++ tests_random/methods/change_nrows.py
