+++ tests_random/methods/delete_columns_array.py
