+++ tests_random/methods/delete_rows_array.py
