+++ tests_random/methods/rbind_self.py
