+++ tests_random/methods/select_rows_array.py
