+++ tests_random/methods/slice_columns.py
