+++ tests_random/methods/slice_rows.py
